+++ src/main/java/ch/admin/bj/swiyu/didtoolbox/securosys/primus/HsmProofOfPossessionJWSSigner.java
@@ -6,6 +6,7 @@ import com.nimbusds.jose.JWSAlgorithm;
 import com.nimbusds.jose.JWSHeader;
 import com.nimbusds.jose.JWSSigner;
 import com.nimbusds.jose.crypto.ECDSASigner;
+import com.nimbusds.jose.jca.JCAContext;
 import com.nimbusds.jose.jwk.ECKey;
 import com.nimbusds.jose.util.Base64URL;
 
@@ -13,7 +14,7 @@ import java.io.IOException;
 import java.security.*;
 import java.security.cert.CertificateException;
 import java.security.interfaces.ECPrivateKey;
-import java.util.Set;
+import java.security.interfaces.EdECPrivateKey;
 
 /**
  * {@link HsmProofOfPossessionJWSSigner} provides multiple constructors for different HSM providers intended to be used with {@link ch.admin.bj.swiyu.didtoolbox.ProofOfPossessionCreator}.
@@ -30,34 +31,45 @@ public final class HsmProofOfPossessionJWSSigner implements ProofOfPossessionJWS
     /**
      * Relies on Securosys Primus HSM cluster as source for signing.
      *
-     * @param primus the HSM cluster
+     * @param primus   the HSM cluster
      * @param password to load the key
-     * @param alias of the key inside the HSM cluster
-     * @param kid of the key inside the JWT
+     * @param alias    of the key inside the HSM cluster
+     * @param kid      of the key inside the JWT
      */
     public static HsmProofOfPossessionJWSSigner newPrimusSigner(PrimusKeyStoreLoader primus, String alias, String password, String kid) throws UnrecoverableEntryException, KeyStoreException, NoSuchAlgorithmException, KeyException, JOSEException {
-        var pk = (ECPrivateKey) primus.loadKeyPair(alias, password).getPrivate();
-        var signer = new ECDSASigner(pk);
-        signer.getJCAContext().setProvider(primus.getKeyStore().getProvider());
-        return new HsmProofOfPossessionJWSSigner(signer, kid);
+        KeyPair kp = primus.loadKeyPair(alias, password);
+        switch (kp.getPrivate()) {
+            case ECPrivateKey ecPrivateKey -> {
+                var signer = new ECDSASigner(ecPrivateKey);
+                signer.getJCAContext().setProvider(primus.getKeyStore().getProvider());
+                return new HsmProofOfPossessionJWSSigner(signer, kid);
+            }
+            case EdECPrivateKey ignored -> {
+                var signer = new PrimusEd25519ProofOfPossessionJWSSignerImpl(kp, kid, primus);
+                return new HsmProofOfPossessionJWSSigner(signer, kid);
+            }
+            default -> {
+                throw new IllegalArgumentException();
+            }
+        }
     }
 
     /**
      * Relies on PKCS11 for signing.
      *
-     * @param cfgPath Path to the configuration file
+     * @param cfgPath        Path to the configuration file
      * @param keystoreSecret
      * @param keyId
-     * @param kid of the key inside the JWT
+     * @param kid            of the key inside the JWT
      * @return
      */
-    public static HsmProofOfPossessionJWSSigner newPkcs11Signer(String cfgPath, String keystoreSecret, String keyId, String kid) throws KeyStoreException, CertificateException, IOException, NoSuchAlgorithmException, JOSEException { // NOPMD UseObjectForClearerAPI
+    public static HsmProofOfPossessionJWSSigner newPkcs11Signer(String cfgPath, String keystoreSecret, String keyId, String kid) throws KeyStoreException, CertificateException, IOException, NoSuchAlgorithmException, JOSEException {
         Provider provider = Security.getProvider("SunPKCS11");
         provider = provider.configure(cfgPath);
         Security.addProvider(provider);
         var hsmKeyStore = KeyStore.getInstance("PKCS11", provider);
         hsmKeyStore.load(null, keystoreSecret.toCharArray());
-        var privateKey =  ECKey.load(hsmKeyStore, keyId, keystoreSecret.toCharArray());
+        var privateKey = ECKey.load(hsmKeyStore, keyId, keystoreSecret.toCharArray());
         var signer = new ECDSASigner(privateKey);
         return new HsmProofOfPossessionJWSSigner(signer, kid);
     }
@@ -78,7 +90,7 @@ public final class HsmProofOfPossessionJWSSigner implements ProofOfPossessionJWS
     }
 
     @Override
-    public Set<JWSAlgorithm> supportedJWSAlgorithms() {
-        return Set.of(this.getAlgorithm());
+    public JCAContext getJCAContext() {
+        return this.signer.getJCAContext();
     }
 }